full users test

+++ src/main/java/ru/freemiumhosting/master/service/UserService.java
@@ -3,6 +3,7 @@ package ru.freemiumhosting.master.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
+import ru.freemiumhosting.master.model.dto.EditUserDto;
 import ru.freemiumhosting.master.model.dto.UserDto;
 import ru.freemiumhosting.master.utils.enums.UserRole;
 import ru.freemiumhosting.master.utils.mappers.UserMapper;
@@ -26,8 +27,7 @@ public class UserService {
     }
 
     public void createUser(UserDto userDto) {
-        userRep.findByNameIgnoreCase(userDto.getName())
-                .orElseThrow(() -> new IllegalStateException("Пользоавтель с таким именем уже занят"));
+        checkUniqueName(userDto.getName());
         User user = userMapper.userDtoToUser(userDto);
         user.setPassword(passwordEncoder.encode(userDto.getPassword()));
         user.setUserRole(UserRole.USER);
@@ -36,17 +36,38 @@ public class UserService {
 
 
     public void deleteUser(Long id) {
+        //TODO delete users projects
+        checkUserWithIdExist(id);
+        checkForAdmin(id);
         userRep.deleteById(id);
     }
 
-    public UserDto editUser(UserDto userDto) {
-        boolean isUserExist = userRep.findById(userDto.getId()).isPresent();
-        if (isUserExist) {
-            User user = userMapper.userDtoToUser(userDto);
-            userRep.save(user);
-            return userMapper.userToUserDto(user);
-        } else {
-            throw new EntityNotFoundException(String.format("Пользователь с id %s не найден", userDto.getId()));
+    public UserDto editUser(EditUserDto userDto) {
+        checkUserWithIdExist(userDto.getId());
+        checkUniqueName(userDto.getName());
+        User user = userMapper.editUserDtoToUser(userDto);
+        userRep.save(user);
+        return userMapper.userToUserDto(user);
+    }
+
+    private void checkForAdmin(Long id) {
+        if (userRep.findById(id).get().getUserRole() == UserRole.ADMIN) {
+            throw new IllegalStateException("Нельзя удалить администратора");
         }
     }
+
+    private void checkUserWithIdExist(Long id) {
+        if (userRep.findById(id).isEmpty()) {
+            throw new EntityNotFoundException(String.format("Пользователь с id %s не найден", id));
+        }
+        ;
+    }
+
+    private void checkUniqueName(String name) {
+        if (userRep.findByNameIgnoreCase(name).isPresent()) {
+            throw new IllegalStateException(String.format("Пользователь с именем %s занят", name));
+        }
+    }
+
+
 }
